auth-service/requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
PyJWT
gunicorn
python-dotenv